+++ compiler-plugin/src/main/java/io/ballerina/stdlib/websubhub/task/AnalyserUtils.java
@@ -40,7 +40,7 @@ import java.util.List;
 import java.util.Optional;
 
 /**
- * {@code ValidatorUtils} contains utility functions required for {@code websubhub:Service} validation.
+ * {@code AnalyserUtils} contains utility functions required for {@code websubhub:Service} validation.
  */
 public final class AnalyserUtils {
     public static void updateContext(SyntaxNodeAnalysisContext context, WebSubHubDiagnosticCodes errorCode,
@@ -56,23 +56,23 @@ public final class AnalyserUtils {
             return ((UnionTypeSymbol) listenerType).memberTypeDescriptors().stream()
                     .filter(typeDescriptor -> typeDescriptor instanceof TypeReferenceTypeSymbol)
                     .map(typeReferenceTypeSymbol -> (TypeReferenceTypeSymbol) typeReferenceTypeSymbol)
-                    .anyMatch(typeReferenceTypeSymbol ->
-                            typeReferenceTypeSymbol.getModule().isPresent()
-                                    && isWebSubHub(typeReferenceTypeSymbol.getModule().get()
-                            ));
+                    .anyMatch(AnalyserUtils::isWebSubHubListenerType);
         }
-
         if (listenerType.typeKind() == TypeDescKind.TYPE_REFERENCE) {
-            Optional<ModuleSymbol> moduleOpt = ((TypeReferenceTypeSymbol) listenerType).typeDescriptor().getModule();
-            return moduleOpt.isPresent() && isWebSubHub(moduleOpt.get());
+            return isWebSubHubListenerType((TypeReferenceTypeSymbol) listenerType);
         }
-
         if (listenerType.typeKind() == TypeDescKind.OBJECT) {
             Optional<ModuleSymbol> moduleOpt = listenerType.getModule();
             return moduleOpt.isPresent() && isWebSubHub(moduleOpt.get());
         }
-
         return false;
+    }
+
+    private static boolean isWebSubHubListenerType(TypeReferenceTypeSymbol typeSymbol) {
+        if (typeSymbol.getName().isEmpty() || !Constants.LISTENER_IDENTIFIER.equals(typeSymbol.getName().get())) {
+            return false;
+        }
+        return typeSymbol.getModule().isPresent() && isWebSubHub(typeSymbol.getModule().get());
     }
 
     public static boolean isWebSubHub(ModuleSymbol moduleSymbol) {